Volume level constantly decreases #25
When we set the volumen to, Ex.: 33 and then we tried to get the current volume
the value returned by osascript is 32 (current_volumen - 1)
Since this is running inside a loop, for every iteration we end up with the flow:

 - Get current volume
 - Set current volume (at this point the new volume is set one point below the actual volume value)

So for every iteration we were decreasing the volumen by 1 causing the volume to drop up to 3 or 4
points every time we pause or a new song is played.

This doesn't happen when the volume is set to 100 for some reason.